requirements.txt
streamlit
openai
python-dotenv
requests
aiohttp
beautifulsoup4
instagrapi
fastapi
uvicorn
pydantic
